requirements.txt
django
pymongo
python-dotenv
